refactor(components): [checkbox]adjust composables folder structure
