+++ packages/components/checkbox/src/composables/index.ts
@@ -1,1 +1,5 @@
+export * from './use-checkbox-disabled'
+export * from './use-checkbox-event'
+export * from './use-checkbox-status'
+export * from './use-checkbox-model'
 export * from './use-checkbox'
